nixos/scx: enable stricter check by hardcoding possible schedulers
Currently we can put any value like "hello-world" there and there is no way to check if that is appropiate.
This checking allows correct values to be used and warns in case of mistake.
This should check for typo errors like "scx-rustland" instead of "scx_rustland" which does not start the service properly.