helix_git: fix broken symlinks for grammars
Based on the the grammar utility in nixpkgs located at nixpkgs/pkgs/development/tools/parsing/tree-sitter/grammar.nix,
which is used by the helix_git package, the output artifact that should be expected is called `parser` and not
`<lang>.so` as is currently the case.

Due to the broken symlink, the Helix editor is unable to locate the tree-sitter grammar when it starts up.

This minor change rectifies the broken symlink by pointing the the correct, expected artifact produced by the nixpkgs
utility i.e. `parser`.